feat: evaluator now handles while loops
