+++ src/interpreter/mod.rs
@@ -1,1 +1,2 @@
 pub mod evaluator;
+mod evaluator_tests;
